Support legacy updates
This also introduces the ability to send in only the attributes you want to update instead of sending all object attributes in the update call.

I'm not sure if this initial pattern is the best because it's not generic and needs to be supported in each models `update_xml` method. I will hopefully revisit this to see if there is a more generic way to only send the attributes that you want to update.

This commit also properly extracts legacy response errors and status.